+++ test/bitfinexIntegrationTests.js
@@ -9,8 +9,10 @@ var apiHost = {
     public:  "https://api.bitfinex.com"
 };
 
-var publicMethodsToTest = ["getTicker", "getRate", "getOrderBook"];
-var privateMethodsToTest = ["getBalance"];
+//var publicMethodsToTest = ["getTicker", "getRate", "getOrderBook"];
+//var privateMethodsToTest = ["getBalance", "getTransactions"];
+var publicMethodsToTest = ["getTicker"];
+var privateMethodsToTest = ["getTransactions"];
 var writeMockResponseFileForMethod = "";
 
 // don't change below this line; only configure above this line