+++ test/bitstampIntegrationTests.js
@@ -10,7 +10,7 @@ var apiHost = {
 };
 
 var publicMethodsToTest = ["getTicker", "getRate", "getOrderBook"];
-var privateMethodsToTest = ["getFee", "getTransactions"];
+var privateMethodsToTest = ["getFee", "getBalance", "getTransactions"];
 var writeMockResponseFileForMethod = "";
 
 // don't change below this line; only configure above this line